moved lineshape functions to new file
