+++ mermithid/misc/__init__.py
@@ -8,3 +8,4 @@ from . import KuriePlotTools
 from . import TritiumFormFactor
 from . import FakeTritiumDataFunctions
 from . import ConversionFunctions
+from . import KrLineshapeFunctions
